Rebased the language settings in settings class and implemented the save by the app of the active language
